separate files for writing output and converters, flatten bedtime json before data frame
