+++ .travis.yml
@@ -1,5 +1,7 @@
 language: python
 python:
+  - 3.5
+  - 3.6
   - 3.7
   - 3.8
 install:
