+++ oura/client_pandas.py
@@ -1,6 +1,7 @@
 import pandas as pd
 
 from .client import OuraClient
+from .converters import ActivityConverter, SleepConverter, UnitConverter
 
 
 class OuraClientDataFrame(OuraClient):
@@ -27,7 +28,7 @@ class OuraClientDataFrame(OuraClient):
             personal_access_token,
         )
 
-    def _summary_df(self, summary, metrics=None):
+    def _summary_df(self, summary, metrics=None, date_key="summary_date"):
         """
         Creates a dataframe from a summary object
 
@@ -51,8 +52,8 @@ class OuraClientDataFrame(OuraClient):
             if "summary_date" not in metrics:
                 metrics.insert(0, "summary_date")
             df = df[metrics]
-        df["summary_date"] = pd.to_datetime(df["summary_date"]).dt.date
-        df = df.set_index("summary_date")
+        df[date_key] = pd.to_datetime(df[date_key]).dt.date
+        df = df.set_index(date_key)
         return df
 
     def sleep_df_raw(self, start=None, end=None, metrics=None):
@@ -157,6 +158,27 @@ class OuraClientDataFrame(OuraClient):
         """
         return self.readiness_df_raw(start, end, metrics)
 
+    def bedtime_df_raw(self, start=None, end=None, metrics=None):
+        """
+        Create a dataframe from bedtime summary
+        The dataframe is minimally edited, i.e 'raw'
+
+        :param start: Beginning of date range
+        :type start: string representation of a date i.e. '2020-10-31'
+
+        :param end: End of date range, or None if you want the current day.
+        :type end: string representation of a date i.e. '2020-10-31'
+
+        :param metrics: Metrics to include in the df.
+        :type metrics: A list of strings, or a string
+        """
+        bedtime_summary = super().bedtime_summary(start, end)["ideal_bedtimes"]
+        for s in bedtime_summary:
+            s["window_start"] = s["bedtime_window"]["start"]
+            s["window_end"] = s["bedtime_window"]["end"]
+            del s["bedtime_window"]
+        return self._summary_df(bedtime_summary, metrics, date_key="date")
+
     def combined_df_edited(self, start=None, end=None, metrics=None):
         """
         Combines sleep, activity, and summary into one DF
@@ -196,163 +218,3 @@ class OuraClientDataFrame(OuraClient):
             activity_df, on="summary_date"
         )
         return combined_df
-
-    def save_as_xlsx(self, df, file, index=True, **to_excel_kwargs):
-        """
-        Save dataframe as .xlsx file with dates properly formatted
-
-        :param df: dataframe to save
-        :type df: df object
-
-        :param file: File path
-        :type file: string
-
-        :param index: save df index, in this case summary_date
-        :type index: Boolean
-        """
-
-        def localize(df):
-            """
-            Remove tz from datetime cols since Excel doesn't allow
-            """
-            tz_cols = df.select_dtypes(include=["datetimetz"]).columns
-            for tz_col in tz_cols:
-                df[tz_col] = df[tz_col].dt.tz_localize(None)
-            return df
-
-        import xlsxwriter
-
-        df = df.copy()
-        df = localize(df)
-        writer = pd.ExcelWriter(
-            file,
-            engine="xlsxwriter",
-            date_format="m/d/yyy",
-            datetime_format="m/d/yyy h:mmAM/PM",
-        )
-        df.to_excel(writer, index=index, **to_excel_kwargs)
-        writer.save()
-
-    def tableize(self, df, tablefmt="pretty", is_print=True, filename=None):
-        """
-        Converts dataframe to a formatted table
-        For more details, see https://pypi.org/project/tabulate/
-
-        :param df: dataframe to save
-        :type df: df object
-
-        :param tablefmt: format of table
-        :type tablefmt: string
-
-        :param is_print: print to standard output?
-        :type is_print: boolean
-
-        :param filename: optionally, filename to print to
-        :type filename: string
-        """
-        from tabulate import tabulate
-
-        table = tabulate(
-            df,
-            headers="keys",
-            tablefmt=tablefmt,
-            showindex=True,
-            stralign="center",
-            numalign="center",
-        )
-        if is_print:
-            print(table)
-        if filename:
-            with open(filename, "w") as f:
-                print(table, file=f)
-        return table
-
-
-class UnitConverter:
-    """
-    Use this class to convert units for certain dataframe cols
-    """
-
-    all_dt_metrics = []
-    all_sec_metrics = []
-
-    def rename_converted_cols(self, df, metrics, suffix_str):
-        """
-        Rename converted cols by adding a suffix to the col name
-        For example, 'bedtime_start' becomes 'bedtime_start_dt_adjusted'
-
-        :param df: a dataframe
-        :type df: pandas dataframe obj
-
-        :param metrics: metrics to rename
-        :type metrics: list of strings
-
-        :param suffix_str: the str to append to each metric name
-        :type suffix_str: str
-        """
-        updated_headers = [header + suffix_str for header in metrics]
-        d_to_rename = dict(zip(metrics, updated_headers))
-        df = df.rename(columns=d_to_rename)
-        return df
-
-    def convert_to_dt(self, df, dt_metrics):
-        """
-        Convert dataframe fields to datetime dtypes
-
-        :param df: dataframe
-        :type df: pandas dataframe obj
-
-        :param dt_metrics: List of metrics to be converted to datetime
-        :type dt_metrics: List
-        """
-        for i, dt_metric in enumerate(dt_metrics):
-            df[dt_metric] = pd.to_datetime(df[dt_metric], format="%Y-%m-%d %H:%M:%S")
-        df = self.rename_converted_cols(df, dt_metrics, "_dt_adjusted")
-        return df
-
-    def convert_to_hrs(self, df, sec_metrics):
-        """
-        Convert fields from seconds to minutes
-
-        :param df: dataframe
-        :type df: pandas dataframe obj
-
-        :param sec_metrics: List of metrics to be converted from sec -> hrs
-        :type sec_metrics: List
-        """
-        df[sec_metrics] = df[sec_metrics] / 60 / 60
-        df = self.rename_converted_cols(df, sec_metrics, "_in_hrs")
-        return df
-
-    def convert_metrics(self, df):
-        """
-        Convert metrics to new unit type
-
-        :param df: dataframe
-        :type df: pandas dataframe obj
-        """
-        dt_metrics = [col for col in df.columns if col in self.all_dt_metrics]
-        sec_metrics = [col for col in df.columns if col in self.all_sec_metrics]
-        if dt_metrics:
-            df = self.convert_to_dt(df, dt_metrics)
-        if sec_metrics:
-            df = self.convert_to_hrs(df, sec_metrics)
-        return df
-
-
-class SleepConverter(UnitConverter):
-    all_dt_metrics = ["bedtime_end", "bedtime_start"]
-    all_sec_metrics = [
-        "awake",
-        "deep",
-        "duration",
-        "light",
-        "onset_latency",
-        "rem",
-        "total",
-    ]
-
-
-class ActivityConverter(UnitConverter):
-    all_dt_metrics = ["day_end", "day_start"]
-    all_sec_metrics = []
